dags/requirements.txt


apache-airflow
apache-airflow-providers-sqlite
apache-airflow-providers-common-sql
pendulum
requests
xmltodict